fastapi-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-pptx==1.0.2
Pillow==10.4.0
httpx==0.27.2
supabase==2.7.4
pydantic-settings==2.4.0
